refactor(html-theme): Extract handleLocalImageLink.ts from getStaticFileHandler.ts
This improves code readability
